PaletteMixer/pyproject.toml
[project]
name = "palettemixer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "color-science>=0.4.5",
    "colour-science>=0.4.7",
    "matplotlib>=3.10.8",
    "pillow>=12.0.0",
    "pymixbox>=2.0.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scipy>=1.16.3",
]
